PR review fixes: clamp tokenizer depth, extract tokenizer, trim docs
Address PR #26 review feedback:

- JsonLimitTokenizer.consumeOutsideString clamps `depth` at zero on stray
  closers (`}` / `]` with depth already 0). Without the clamp, a payload like
  `}}{{{...` would drive `depth` negative on the leading closers and let the
  in-flight peak reach `maxDepth + leadingClosers` before the limit fires.
  kotlinx rejects the mismatched JSON downstream and the entry-size cap bounds
  it in practice, but the invariant `0 <= depth <= maxDepth` is cheap to hold
  and removes any reliance on those downstream bounds. Regression test added
  (`strayLeadingClosersDoNotInflateInFlightDepth`).

- Move the tokenizer (`enforceJsonLimits` + `JsonLimitTokenizer` +
  byte-constants) to its own file, `JsonLimitTokenizer.kt`. Leaves
  PassJsonDecoder.kt as purely the structural-mapping layer.

- Trim the routing-table duplication in PassJsonDecodeResult.kt: keep one
  canonical statement of the parser-glue routing intent at the type-level
  doc, shrink each arm's KDoc to one line. PassJsonDecoder.kt's three-layer
  defense-in-depth prose is the load-bearing routing description.

- Restructure `assemblePass` around a `RequiredFields` extraction step. Drops
  the apology comment about detekt complexity thresholds — the structure now
  reads as the natural one (parse expiration, read requireds, build).

Caveats noted in the review filed as follow-up beads:

- wpass-k5t (P3): naked-local timestamp acceptance for expirationDate.
- wpass-rj7 (P4): nullable `version` on UnknownFormatVersion to distinguish
  missing from non-1 without a sentinel.
